pyproject.toml
[project]
name = "nju-deepseek"
version = "0.3.1"
description = "A Python package for accessing agents listed on https://chat.nju.edu.cn/deepseek via command-line interface."
authors = [
    {name = "Nemowang",email = "211108100@smail.nju.edu.cn"}
]
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
    "python-socketio[client]>=5.13.0",
    "requests>=2.32.3",
    "tenacity>=9.1.2",
    "websocket-client>=1.8.0",
]

[project.scripts]
chat = "nju_deepseek.cli:main"

[project.optional-dependencies]
auth = [
    "lxml>=5.4.0",
    "onnxruntime>=1.22.0",
    "pillow>=11.2.1",
    "pycryptodomex>=3.23.0",
]
cli = [
    "platformdirs>=4.3.8",
    "prompt-toolkit>=3.0.51",
]


[build-system]
requires = ["uv_build>=0.7.8,<0.8.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""
